fix header file betty error

+++ sort.h
@@ -27,7 +27,7 @@ void selection_sort(int *array, size_t size);
 void swap(int *a, int *b);
 void quick_sort(int *array, size_t size);
 void sort_arr(int *array, unsigned int low, unsigned int high, size_t size);
-unsigned int part(int *array, unsigned int low, unsigned int high, size_t size);
+unsigned int part(int *array, size_t low, size_t high, size_t size);
 void shell_sort(int *array, size_t size);
 void counting_sort(int *array, size_t size);
